add getNextSurveyStep to sdk
